edit file register.jsx change into atomic component
